Extract select to own component
