+++ app/routes/queries/listings/index.tsx
@@ -7,11 +7,10 @@ import GetListingRequest from '~/requests/GetListing'
 import type { GetListingProps } from '~/requests/GetListing'
 import NoResults from '~/routes/queries/listings/NoResults'
 import Results from '~/routes/queries/listings/Results'
-import { SubmitButton } from '~/components/form/SubmitButton'
 import { useState } from 'react'
-import { searchForItemName } from '~/utils/items'
 import { getUserSessionData } from '~/sessions'
 import { Differences } from './Differences'
+import { SearchForItem } from './SearchForItem'
 
 const validateInput = ({
   itemId,
@@ -82,7 +81,6 @@ export const ErrorBoundary: ErrorBoundaryComponent = ({ error }) => {
 
 const Index = () => {
   const [id, setId] = useState<number | undefined>()
-  const [name, setName] = useState('')
   const transition = useTransition()
   const results = useActionData()
 
@@ -92,10 +90,6 @@ const Index = () => {
     }
   }
 
-  const items = searchForItemName(name)?.sort()
-
-  const selectIsDisabled = !name || name.length < 2
-
   return (
     <main className="flex-1">
       <div className="py-3">
@@ -104,70 +98,18 @@ const Index = () => {
             <h1 className="text-2xl font-semibold text-blue-900 py-2">
               Get Item Listing Details
             </h1>
-            <div className="mt-3 md:mt-0 md:col-span-3 py-3">
-              <div className="shadow overflow-hidden sm:rounded-md">
-                <div className="px-4 py-2 bg-white sm:p-4">
-                  <div className="flex-1 min-w-full dir-col md:max-w-md">
-                    <div className="col-span-6 sm:col-span-3 xl:col-span-2">
-                      <label
-                        htmlFor="itemName"
-                        className="block text-sm font-medium text-gray-700">
-                        Search for Item by Name
-                      </label>
-                      <div className={`mt-1 flex rounded-md shadow-sm`}>
-                        <input
-                          type={'text'}
-                          id="itemName"
-                          value={name}
-                          placeholder="Potion ..."
-                          onChange={(e) => setName(e.target.value)}
-                          className="flex-1 min-w-0 block w-full px-3 py-2 rounded-none rounded-l-md focus:ring-blue-500 focus:border-blue-500 block w-full shadow-sm sm:text-sm border-gray-300 rounded-md"
-                        />
-                        <span
-                          className={`inline-flex items-center px-3 rounded-r-md border border-l-0 border-gray-300 bg-gray-50 text-gray-500 sm:text-sm`}>
-                          Item
-                        </span>
-                      </div>
-                    </div>
-                  </div>
-                  <div className="max-w-7xl mt-1 flex rounded-md shadow-sm`">
-                    <select
-                      className="flex-1 min-w-0 block w-full px-3 py-2 rounded-md focus:ring-blue-500 focus:border-blue-500 block w-full shadow-sm sm:text-sm border-gray-300 rounded-md"
-                      defaultValue={''}
-                      disabled={selectIsDisabled}
-                      onChange={(e) => setId(parseInt(e.target.value))}>
-                      <option value={''} disabled>
-                        {selectIsDisabled ? '(nothing found)' : 'Choose item'}
-                      </option>
-                      {items &&
-                        items.map(([id, item]) => (
-                          <option key={item} value={id}>
-                            {item}
-                          </option>
-                        ))}
-                    </select>
-                  </div>
-                </div>
-
-                <div className="px-4 py-2 bg-white sm:p-2">
-                  <div className="flex justify-between">
-                    {
-                      <p className="text-red-500 mx-2">
-                        {results && 'exception' in results
-                          ? `Server Error: ${results.exception}`
-                          : ''}
-                      </p>
-                    }
-                    <SubmitButton
-                      title="Search"
-                      onClick={onSubmit}
-                      loading={transition.state === 'submitting'}
-                      disabled={!id}
-                    />
-                  </div>
-                </div>
-              </div>
-            </div>
+            <SearchForItem
+              loading={transition.state === 'submitting'}
+              onClick={onSubmit}
+              onSelectChange={({ id }) => {
+                setId(id)
+              }}
+              error={
+                results && 'exception' in results
+                  ? `Server Error: ${results.exception}`
+                  : undefined
+              }
+            />
           </div>
           <input name="itemId" value={id} hidden />
         </Form>
